third_party/WebKit/LayoutTests/fast/html/tooltip-close.html

<!DOCTYPE html>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id=log></div>
<div id=start title=tooltip>foo</div>
<script>
var div = document.querySelector('#start');
test(function() {
    eventSender.mouseMoveTo(div.offsetLeft, div.offsetTop + div.offsetHeight / 2);
    assert_equals(testRunner.tooltipText, 'tooltip');
    eventSender.keyDown('a');
    assert_equals(testRunner.tooltipText, '');
}, 'Key events should close tooltips.');
</script>
</body>
